fix(email): stop importing conftest as a module across test roots
CI (test_email_agent.yml) runs pytest over six roots in one command;
under prepend import mode each root's directory lands on sys.path, so
the bare top-level name `conftest` is ambiguous across them and
whichever same-named conftest.py sys.path resolves first wins. Three
test modules did `from conftest import FakeAgent, ScriptedConsole`,
which broke collection entirely under the real multi-root invocation
(ImportError: cannot import name 'FakeAgent' from 'conftest') even
though it passed locally against the narrower two-root run.

Move the two fakes into a uniquely-named onboarding_fakes.py instead —
conftest.py itself is fine (pytest discovers it per-directory, never
imported by name); only an explicit cross-file import of it collides.

Verified against CI's actual invocation, not the narrower local one:
python -m pytest hub/agents/email/python/tests/ \
  tests/test_email_openapi_conformance.py tests/unit/agents/email/ \
  tests/unit/email/ tests/unit/agents/test_email_agent.py \
  tests/unit/agents/test_email_agent_confirmation.py -q
-> 2031 passed, 1 skipped.
